+++ src/Client.php
@@ -304,9 +304,7 @@ class Client implements ClientInterface, \Psr\Http\Client\ClientInterface
         $request = $this->applyOptions($request, $options);
 
         if ('' === $request->getProtocolVersion()) {
-            trigger_deprecation('guzzlehttp/guzzle', '7.11', 'Sending a request with an empty protocol version is deprecated; guzzlehttp/guzzle 8.0 will reject empty protocol versions.');
-
-            $request = Psr7\Utils::modifyRequest($request, ['version' => '1.1']);
+            throw new InvalidArgumentException('HTTP protocol version must not be empty.');
         }
 
         /** @var HandlerStack $handler */
@@ -479,9 +477,7 @@ class Client implements ClientInterface, \Psr\Http\Client\ClientInterface
     private static function normalizeProtocolVersion($version): string
     {
         if ('' === $version) {
-            trigger_deprecation('guzzlehttp/guzzle', '7.11', 'Passing an empty "version" request option is deprecated; guzzlehttp/guzzle 8.0 will reject empty protocol versions.');
-
-            return '1.1';
+            throw new InvalidArgumentException('HTTP protocol version must not be empty.');
         }
 
         return \is_float($version) ? \number_format($version, 1, '.', '') : (string) $version;